requirements.txt
django
djangorestframework
djangorestframework-bulk
django-all-access
django-widget-tweaks
